+++ cmd/neofs-cli/modules/object/get.go
@@ -10,9 +10,10 @@ import (
 	internalclient "github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/client"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/commonflags"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/key"
+	"github.com/nspcc-dev/neofs-sdk-go/client"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
-	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
+	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/spf13/cobra"
 )
 
@@ -46,7 +47,7 @@ func getObject(cmd *cobra.Command, _ []string) error {
 	var cnr cid.ID
 	var obj oid.ID
 
-	objAddr, err := readObjectAddress(cmd, &cnr, &obj)
+	_, err := readObjectAddress(cmd, &cnr, &obj)
 	if err != nil {
 		return err
 	}
@@ -80,9 +81,7 @@ func getObject(cmd *cobra.Command, _ []string) error {
 	}
 	defer cli.Close()
 
-	var prm internalclient.GetObjectPrm
-	prm.SetClient(cli)
-	prm.SetPrivateKey(*pk)
+	var prm client.PrmObjectGet
 	err = Prepare(cmd, &prm)
 	if err != nil {
 		return err
@@ -93,8 +92,9 @@ func getObject(cmd *cobra.Command, _ []string) error {
 	}
 
 	raw, _ := cmd.Flags().GetBool(rawFlag)
-	prm.SetRawFlag(raw)
-	prm.SetAddress(objAddr)
+	if raw {
+		prm.MarkRaw()
+	}
 
 	var p *pb.ProgressBar
 	noProgress, _ := cmd.Flags().GetBool(noProgressFlag)
@@ -109,19 +109,21 @@ func getObject(cmd *cobra.Command, _ []string) error {
 		payloadWriter = out
 	}
 
-	if filename == "" || noProgress {
-		prm.SetPayloadWriter(payloadWriter)
-	} else {
-		p = pb.New64(0)
-		p.Output = cmd.OutOrStdout()
-		prm.SetPayloadWriter(p.NewProxyWriter(payloadWriter))
-		prm.SetHeaderCallback(func(o *object.Object) {
-			p.SetTotal64(int64(o.PayloadSize()))
+	hdr, rdr, err := cli.ObjectGetInit(ctx, cnr, obj, user.NewAutoIDSigner(*pk), prm)
+	if err == nil {
+		if filename != "" && !noProgress {
+			p = pb.New64(0)
+			p.Output = cmd.OutOrStdout()
+			p.SetTotal64(int64(hdr.PayloadSize()))
 			p.Start()
-		})
-	}
 
-	res, err := internalclient.GetObject(ctx, prm)
+			payloadWriter = p.NewProxyWriter(payloadWriter)
+		}
+
+		if _, err = io.Copy(payloadWriter, rdr); err != nil {
+			err = fmt.Errorf("copy payload: %w", err)
+		}
+	}
 	if p != nil {
 		p.Finish()
 	}
@@ -135,7 +137,7 @@ func getObject(cmd *cobra.Command, _ []string) error {
 	}
 
 	if binary {
-		objToStore := res.Header()
+		objToStore := hdr
 		// TODO(@acid-ant): #1932 Use streams to marshal/unmarshal payload
 		objToStore.SetPayload(payloadBuffer.Bytes())
 		objBytes := objToStore.Marshal()
@@ -151,7 +153,7 @@ func getObject(cmd *cobra.Command, _ []string) error {
 
 	// Print header only if file is not streamed to stdout.
 	if filename != "" {
-		err = printHeader(cmd, res.Header())
+		err = printHeader(cmd, &hdr)
 		if err != nil {
 			return err
 		}